fix(dht): set pub visibility for reqres response data

+++ market_dht/src/behaviour/file_req_res.rs
@@ -127,6 +127,6 @@ pub(crate) struct FileMetadata {
 pub struct SupplierInfo {
     pub(crate) ip: Ipv4Addr,
     pub(crate) port: u16,
-    pub(crate) price: i32,
+    pub(crate) price: i64,
     pub(crate) username: String,
 }
